Extract CasperPadketHandlerInternal into separate file
